pyproject.toml
[tool.poetry]
name = "image-to-webp"
version = "0.1.0"
description = ""
authors = ["Caner Tuncay <caner.tuncay@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flet = "^0.19.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
